Remove re2j regexp engine support
