+++ core/trino-main/pom.xml
@@ -300,11 +300,6 @@
 
         <dependency>
             <groupId>io.trino</groupId>
-            <artifactId>trino-re2j</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>io.trino</groupId>
             <artifactId>trino-spi</artifactId>
         </dependency>
