+++ core/trino-main/src/main/java/io/trino/FeaturesConfig.java
@@ -23,7 +23,6 @@ import io.airlift.configuration.LegacyConfig;
 import io.airlift.units.DataSize;
 import io.airlift.units.MaxDataSize;
 import io.trino.execution.buffer.CompressionCodec;
-import io.trino.sql.analyzer.RegexLibrary;
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.Min;
@@ -38,7 +37,6 @@ import static io.airlift.units.DataSize.Unit.KILOBYTE;
 import static io.airlift.units.DataSize.succinctBytes;
 import static io.trino.execution.buffer.CompressionCodec.LZ4;
 import static io.trino.execution.buffer.CompressionCodec.NONE;
-import static io.trino.sql.analyzer.RegexLibrary.JONI;
 
 @DefunctConfig({
         "analyzer.experimental-syntax-enabled",
@@ -68,7 +66,10 @@ import static io.trino.sql.analyzer.RegexLibrary.JONI;
         "experimental.spill-window-operator",
         "legacy.allow-set-view-authorization",
         "parse-decimal-literals-as-double",
-        "experimental.late-materialization.enabled"
+        "experimental.late-materialization.enabled",
+        "regex-library",
+        "re2j.dfa-states-limit",
+        "re2j.dfa-retries"
 })
 public class FeaturesConfig
 {
@@ -88,9 +89,6 @@ public class FeaturesConfig
     private boolean omitDateTimeTypePrecision;
     private int maxRecursionDepth = 10;
 
-    private int re2JDfaStatesLimit = Integer.MAX_VALUE;
-    private int re2JDfaRetries = 5;
-    private RegexLibrary regexLibrary = JONI;
     private boolean spillEnabled;
     private DataSize aggregationOperatorUnspillMemoryLimit = DataSize.of(4, DataSize.Unit.MEGABYTE);
     private List<Path> spillerSpillPaths = ImmutableList.of();
@@ -191,44 +189,6 @@ public class FeaturesConfig
     public FeaturesConfig setMaxMemoryPerPartitionWriter(DataSize maxMemoryPerPartitionWriter)
     {
         this.maxMemoryPerPartitionWriter = maxMemoryPerPartitionWriter;
-        return this;
-    }
-
-    @Min(2)
-    public int getRe2JDfaStatesLimit()
-    {
-        return re2JDfaStatesLimit;
-    }
-
-    @Config("re2j.dfa-states-limit")
-    public FeaturesConfig setRe2JDfaStatesLimit(int re2JDfaStatesLimit)
-    {
-        this.re2JDfaStatesLimit = re2JDfaStatesLimit;
-        return this;
-    }
-
-    @Min(0)
-    public int getRe2JDfaRetries()
-    {
-        return re2JDfaRetries;
-    }
-
-    @Config("re2j.dfa-retries")
-    public FeaturesConfig setRe2JDfaRetries(int re2JDfaRetries)
-    {
-        this.re2JDfaRetries = re2JDfaRetries;
-        return this;
-    }
-
-    public RegexLibrary getRegexLibrary()
-    {
-        return regexLibrary;
-    }
-
-    @Config("regex-library")
-    public FeaturesConfig setRegexLibrary(RegexLibrary regexLibrary)
-    {
-        this.regexLibrary = regexLibrary;
         return this;
     }
